+++ spec/Akeneo/Bundle/RuleEngineBundle/EventSubscriber/LogExecutionSubscriberSpec.php
@@ -2,12 +2,12 @@
 
 namespace spec\Akeneo\Bundle\RuleEngineBundle\EventSubscriber;
 
-use Akeneo\Bundle\RuleEngineBundle\Event\SkippedSubjectRuleEvent;
-use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Akeneo\Bundle\RuleEngineBundle\Event\SelectedRuleEvent;
+use Akeneo\Bundle\RuleEngineBundle\Event\SkippedSubjectRuleEvent;
 use Akeneo\Bundle\RuleEngineBundle\Model\RuleDefinitionInterface;
 use Akeneo\Bundle\RuleEngineBundle\Model\RuleSubjectSetInterface;
+use PhpSpec\ObjectBehavior;
+use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Prophecy\Argument;
 use Psr\Log\LoggerInterface;