+++ spec/PimEnterprise/Bundle/EnrichBundle/MassEditAction/Operation/PublishSpec.php
@@ -3,34 +3,54 @@
 namespace spec\PimEnterprise\Bundle\EnrichBundle\MassEditAction\Operation;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Model\ProductInterface;
-use PimEnterprise\Bundle\SecurityBundle\Attributes;
-use PimEnterprise\Bundle\WorkflowBundle\Manager\PublishedProductManager;
-use Prophecy\Argument;
-use Symfony\Component\Security\Core\SecurityContextInterface;
 
 class PublishSpec extends ObjectBehavior
 {
-    function it_is_initializable()
+    function it_is_a_mass_edit_operation()
     {
-        $this->shouldHaveType('PimEnterprise\Bundle\EnrichBundle\MassEditAction\Operation\Publish');
+        $this->shouldImplement('Pim\Bundle\EnrichBundle\MassEditAction\Operation\MassEditOperationInterface');
+        $this->shouldImplement('Pim\Bundle\EnrichBundle\MassEditAction\Operation\ConfigurableOperationInterface');
+        $this->shouldImplement('Pim\Bundle\EnrichBundle\MassEditAction\Operation\BatchableOperationInterface');
     }
 
-    function let(PublishedProductManager $manager, SecurityContextInterface $securityContext)
-    {
-        $this->beConstructedWith($manager, $securityContext);
-    }
-
-    function it_gets_form_type_alias()
+    function it_provides_a_form_type()
     {
         $this->getFormType()->shouldReturn('pimee_enrich_mass_publish');
     }
 
-    function it_publishes_each_product(ProductInterface $foo, ProductInterface $bar, $securityContext, $manager)
+    function it_provides_form_options()
     {
-        $this->setObjectsToMassEdit([$foo, $bar]);
-        $securityContext->isGranted(Attributes::OWN, Argument::any())->willReturn(true);
-        $manager->publishAll([$foo, $bar])->shouldBeCalled();
-        $this->perform();
+        $this->getFormOptions()->shouldReturn([]);
+    }
+
+    function it_provides_items_name_it_works_on()
+    {
+        $this->getItemsName()->shouldReturn('product');
+    }
+
+    function it_provides_an_alias()
+    {
+        $this->getOperationAlias()->shouldReturn('publish');
+    }
+
+    function it_provides_correct_actions_to_apply_on_products()
+    {
+        $this->getActions()->shouldReturn([]);
+    }
+
+    function it_provides_a_batch_job_code()
+    {
+        $this->getBatchJobCode()->shouldReturn('publish_product');
+    }
+
+    function it_provides_formatted_batch_config_for_the_job()
+    {
+        $this->setFilters([
+            ['id', 'IN', ['100', '50']]
+        ]);
+
+        $this->getBatchConfig()->shouldReturn(
+            '{\"filters\":[[\"id\",\"IN\",[\"100\",\"50\"]]],\"actions\":[]}'
+        );
     }
 }