+++ src/Akeneo/Bundle/RuleEngineBundle/Model/ActionInterface.php
@@ -16,5 +16,4 @@ namespace Akeneo\Bundle\RuleEngineBundle\Model;
  */
 interface ActionInterface
 {
-
 }