+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+# 1.4.x
+
+## Bug fixes
+- PIM-3933: Add missing translation keys for published products
+- PIM-3296: Better display of options changes in Proposals
+
+## BC Breaks
+- Remove `PimEnterprise\Bundle\WorkflowBundle\Comparator\DateComparator`
+- Rename `PimEnterprise\Bundle\WorkflowBundle\Comparator\ComparatorInterface` to `PimEnterprise\Bundle\WorkflowBundle\Comparator\ComparatorInterface`. First argument of method `supportsComparison` is replaced by a string $type
+- `PimEnterprise\Bundle\WorkflowBundle\Comparator\ChainedComparator` is replaced by `PimEnterprise\Bundle\WorkflowBundle\Comparator\ComparatorRegistry`
+- Replace tag `pimee_workflow.comparator` by `pimee_workflow.attribute.comparator`
+- Remove `PimEnterprise\Bundle\WorkflowBundle\Doctrine\Common\Saver\ProductDraftSaver`
+- Remove `PimEnterprise\Bundle\WorkflowBundle\Event\ChangeSetEvent`
+- Remove `PimEnterprise\Bundle\WorkflowBundle\Event\ChangeSetEvents`
+- Remove `PimEnterprise\Bundle\WorkflowBundle\Event\ProductDraftEvent`
+- Remove `PimEnterprise\Bundle\WorkflowBundle\EventSubscriber\ChangeSet\MetadataSubscriber`
+- Remove `PimEnterprise\Bundle\WorkflowBundle\EventSubscriber\ProductDraft\MarkInProgressSubscriber`
+- Remove `PimEnterprise\Bundle\WorkflowBundle\EventSubscriber\ProductDraft\PrepareProductDraftChangesSubscriber`
+- Remove `PimEnterprise\Bundle\WorkflowBundle\EventSubscriber\ProductDraft\PrepareUploadingMediaSubscriber`
+- Remove `PimEnterprise\Bundle\WorkflowBundle\EventSubscriber\ProductDraft\UpdateProductDraftStatusSubscriber`
+- Change the constructor of `PimEnterprise\Bundle\WorkflowBundle\Manager\ProductDraftManager`, removed the first argument $registry, added `Akeneo\Component\StorageUtils\Saver\SaverInterface` and `Akeneo\Component\StorageUtils\Remover\RemoverInterface` as the latest arguments
+- Remove second argument of `supports` method of `PimEnterprise\Bundle\WorkflowBundle\Presenter\PresenterInterface`
+- Move PimEnterprise/Bundle/CatalogBundle/Doctrine/MongoDBODM/{ → Repository}/ProductMassActionRepository
+- Move PimEnterprise/Bundle/CatalogBundle/Doctrine/ORM/{ → Repository}/ProductMassActionRepository
+- Move PimEnterprise/Bundle/CatalogBundle/{Entity → Doctrine/ORM}/Repository/AttributeRepository
+- Move PimEnterprise/Bundle/WorkflowBundle/Doctrine/MongoDBODM/{ → Repository}/ProductDraftRepository
+- Move PimEnterprise/Bundle/WorkflowBundle/Doctrine/MongoDBODM/{ → Repository}/PublishedProductRepository
+- Move PimEnterprise/Bundle/WorkflowBundle/Doctrine/ORM/{ → Repository}/ProductDraftRepository
+- Move PimEnterprise/Bundle/WorkflowBundle/Doctrine/ORM/{ → Repository}/PublishedAssociationRepository
+- Move PimEnterprise/Bundle/WorkflowBundle/Doctrine/ORM/{ → Repository}/PublishedProductRepository
+- Add ProductBuilderInterface argument of the constructor of PimEnterprise/Bundle/CatalogRuleBundle/Validator/Constraints/ProductRule/ValueActionValidator
+
 # 1.3.x
 
 ## Bug fixes
@@ -10,6 +42,8 @@
 - PIM-4227: Disable product versionning on category update (never used and very slow)
 
 # 1.3.12 (2015-05-22)
+
+# 1.3.9 (2015-04-21)
 
 # 1.3.7 (2015-04-03)
 
@@ -51,9 +85,17 @@
 
 ## Bug fixes
 - PIM-3760: Fix reverting previous product versions with file/image attributes
+- PIM-3784: Generate the completeness on selected products before to mass-publish
 - PIM-3765: Do not allow the revert of a product if it has a variant group
 
+## BC breaks
+- Update constructor of `PimEnterprise/Bundle/WorkflowBundle/Publisher/Product/ProductPublisher` to add a `Pim\Bundle\CatalogBundle\Manager\CompletenessManager` $completenessManager argument
+- Update constructor of `PimEnterprise/Bundle/VersioningBundle/Reverter/ProductReverter` to add a `Symfony\Component\Translation\TranslatorInterface` $translator argument
+
 # 1.3.0 - "Strawberry" (2015-02-12)
+
+## Bug fixes
+- PIM-3760: Fix reverting previous product versions with file/image attributes
 
 # 1.3.0-RC2 (2015-02-12)